feat: add unit tests with mutation for Actions

+++ app/Actions/Pages/MentorProgram/DestroyMentorProgramPage.php
@@ -5,15 +5,20 @@ declare(strict_types=1);
 namespace App\Actions\Pages\MentorProgram;
 
 use App\Models\MentorProgram;
+use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Http\RedirectResponse;
+use Illuminate\Http\Request;
 use Lorisleiva\Actions\Concerns\AsController;
 
 class DestroyMentorProgramPage
 {
     use AsController;
 
-    public function handle(MentorProgram $mentorProgram): RedirectResponse
+    public function handle(Request $request, MentorProgram $mentorProgram): RedirectResponse
     {
+        throw_unless($mentorProgram->exists, new ModelNotFoundException('Mentor program not found.'));
+
+        abort_if($mentorProgram->mentor_id !== $request->user()->id, 403, 'Unauthorized action.');
         $mentorProgram->delete();
 
         return redirect()->route('mentor-program.create');
